move to net module and fix madsim
